(cleanup) Reorganize folder structure
