+++ controllers/userController.js
@@ -3,7 +3,7 @@ var User = require('../models/user.js');
 
 module.exports.saveOne = function(req, res) {
   console.log('Saved one user!');
-
+  res.json(1);
   // EXAMPLE CODE:
   // var user = req.body;
   // User.findOrCreate({where: user})
